Add depDouble support to multinomPois. Update manual. Consolidate C++ code for pifuns.
