+++ src/nll_multinomPois.cpp
@@ -1,4 +1,5 @@
 #include "nll_multinomPois.h"
+#include "pifun.h"
 
 using namespace Rcpp;
 using namespace arma;
@@ -6,35 +7,6 @@ using namespace arma;
 mat inv_logit_( mat inp ){
   return(1 / (1 + exp(-1 * inp)));
 }
-
-vec removalPiFun_( vec p ){
-  int J = p.size();
-  vec pi(J);
-  pi(0) = p(0);
-  for(int j=1; j<J; j++){
-    pi(j) = pi(j-1) / p(j-1) * (1-p(j-1)) * p(j);
-  }
-  return(pi);
-}
-
-vec doublePiFun_( vec p ){
-  //p must have 2 columns
-  vec pi(3);
-  pi(0) = p(0) * (1 - p(1));
-  pi(1) = p(1) * (1 - p(0));
-  pi(2) = p(0) * p(1);
-  return(pi);
-}
-
-
-vec piFun_( vec p , std::string pi_fun ){
-  if(pi_fun == "removalPiFun"){
-    return(removalPiFun_(p));
-  } else if(pi_fun == "doublePiFun"){
-    return(doublePiFun_(p));
-  }
-}
-
 
 SEXP nll_multinomPois(SEXP betaR, SEXP pi_funR, 
     SEXP XlamR, SEXP Xlam_offsetR, SEXP XdetR, SEXP Xdet_offsetR,  
@@ -74,7 +46,7 @@ SEXP nll_multinomPois(SEXP betaR, SEXP pi_funR,
     vec na_sub = navec.subvec(y_ind, y_stop); 
     if( ! all(na_sub) ){
 
-      vec pi_lam = piFun_( p.subvec(p_ind, p_stop), pi_fun ) * lambda(m);  
+      vec pi_lam = piFun( p.subvec(p_ind, p_stop), pi_fun ) * lambda(m);  
       
       for (int r=0; r<R; r++){
         if( ! na_sub(r) ){
