+++ src/pifun.h
@@ -1,0 +1,8 @@
+#ifndef _unmarked_PIFUN_H
+#define _unmarked_PIFUN_H
+
+#include <RcppArmadillo.h>
+
+arma::vec piFun( arma::vec p, std::string pi_fun );
+
+#endif
